2720:   format string protection/clean up (CVE-2007-1463, CVE-2007-1464)

+++ src/debug/logger.cpp
@@ -116,7 +116,7 @@ static void set_category_mask(bool * const mask, char const *filter) {
                 }
             }
             if (!iter->name) {
-                g_warning("Unknown debugging category %*s", end - start, start);
+                g_warning("Unknown debugging category %*s", (int)(end - start), start);
             }
         }
         if (*end) {
